+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/facade/SoraVideoChannel.kt
@@ -33,6 +33,7 @@ class SoraVideoChannel(
         var         videoBitrate:      Int? = null,
         private var needLocalRenderer: Boolean = true,
         private var audioEnabled:      Boolean = true,
+        private var sdpSemantics:      PeerConnection.SdpSemantics = PeerConnection.SdpSemantics.PLAN_B,
         private var capturerFactory:   CameraVideoCapturerFactory = DefaultCameraVideoCapturerFactory(context),
         private var listener:          Listener?
 ) {
@@ -195,6 +196,7 @@ class SoraVideoChannel(
             videoCodec   = this@SoraVideoChannel.videoCodec
             audioCodec   = this@SoraVideoChannel.audioCodec
             videoBitrate = this@SoraVideoChannel.videoBitrate
+            sdpSemantics = this@SoraVideoChannel.sdpSemantics
         }
 
         mediaChannel = SoraMediaChannel(